fix: :pencil2: rename folder Pages to pages
