+++ template/pages/ForgotPassword/index.js
